Fix grammar error and issue in media trashed job handler

+++ src/Enterspeed.Source.UmbracoCms/Handlers/Media/EnterspeedMediaTrashedJobHandler.cs
@@ -1,11 +1,10 @@
 ﻿using System.Net;
 using Enterspeed.Source.Sdk.Api.Services;
 using Enterspeed.Source.UmbracoCms.Data.Models;
-using Enterspeed.Source.UmbracoCms.Handlers;
 using Enterspeed.Source.UmbracoCms.Models;
 using Enterspeed.Source.UmbracoCms.Exceptions;
 using Enterspeed.Source.UmbracoCms.Providers;
-using Umbraco.Cms.Core.Services;
+using Microsoft.Extensions.Logging;
 
 namespace Enterspeed.Source.UmbracoCms.Handlers.Media
 {
@@ -13,17 +12,17 @@ namespace Enterspeed.Source.UmbracoCms.Handlers.Media
     {
         private readonly IEnterspeedIngestService _enterspeedIngestService;
         private readonly IEnterspeedConnectionProvider _enterspeedConnectionProvider;
-        private readonly IMediaService _mediaService;
+        private readonly ILogger<EnterspeedMediaTrashedJobHandler> _logger;
 
         public EnterspeedMediaTrashedJobHandler
         (
             IEnterspeedIngestService enterspeedIngestService,
             IEnterspeedConnectionProvider enterspeedConnectionProvider,
-            IMediaService mediaService)
+            ILogger<EnterspeedMediaTrashedJobHandler> logger)
         {
             _enterspeedIngestService = enterspeedIngestService;
             _enterspeedConnectionProvider = enterspeedConnectionProvider;
-            _mediaService = mediaService;
+            _logger = logger;
         }
 
         public bool CanHandle(EnterspeedJob job)
@@ -38,12 +37,18 @@ namespace Enterspeed.Source.UmbracoCms.Handlers.Media
         public void Handle(EnterspeedJob job)
         {
             var parsed = int.TryParse(job.EntityId, out var parsedId);
-            var media = parsed ? _mediaService.GetById(parsedId) : null;
 
-            var deleteResponse = _enterspeedIngestService.Delete(media?.Id.ToString(), _enterspeedConnectionProvider.GetConnection(ConnectionType.Publish));
-            if (!deleteResponse.Success && deleteResponse.Status != HttpStatusCode.NotFound)
+            if (parsed)
             {
-                throw new JobHandlingException($"Failed deleting entity ({job.EntityId}/{job.Culture}). Message: {deleteResponse.Message}");
+                var deleteResponse = _enterspeedIngestService.Delete(media?.Id.ToString(), _enterspeedConnectionProvider.GetConnection(ConnectionType.Publish));
+                if (!deleteResponse.Success && deleteResponse.Status != HttpStatusCode.NotFound)
+                {
+                    throw new JobHandlingException($"Failed deleting entity ({job.EntityId}/{job.Culture}). Message: {deleteResponse.Message}");
+                }
+            }
+            else
+            {
+                _logger.LogWarning("Job.EntityId is not a valid ID");
             }
         }
     }
